adicionando CartUI

+++ GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -5,7 +5,7 @@ namespace GeekShopping.Web.Models
         public long Id {get;set;}
         public string UserId {get;set;}
         public string CouponCode {get;set;}
-        public double PurchaseAmout {get;set;}
+        public decimal PurchaseAmout {get;set;}
         
     }
 }
